Update installation instructions so that they work on Windows
pip based installs of spatial packages don't work seamlessly on Windows, so
install them first using conda and then install DeepForest via pip. This is
only an issue due to the fact that we can't currently get the Python deepforest
package to compile properly for conda due to upstream version issues with pytorch
on conda-forge. Once that is fixed we can shift to installing everything via conda.